requirements.txt
black
streamlit
streamlit-option-menu
mysql-connector-python
python-dotenv
selenium
webdriver_manager
beautifulsoup4